refactor: share getVirtualRowStart across sidebar drop modules
Extract the duplicated getVirtualRowStart helper into sidebar-virtual-row-offset.ts
and import it in both project-header-drop.ts and group-header-drop.ts. Also simplify
the redundant 'parentGroupId' in guard in getSiblingGroupIdsByParent since ProjectGroup
always has parentGroupId.
