+++ src/renderer/src/components/sidebar/group-header-drop.ts
@@ -2,6 +2,7 @@
 import { interpolateSparseOrder } from './sidebar-drop-order-interpolation'
 import { mapSidebarProjectHeaderDropIndexToSiblingInsertIndex } from './project-header-drop'
 import { getWorktreeSidebarBoundaryDrop } from './worktree-sidebar-drag-autoscroll'
+import { getVirtualRowStart } from './sidebar-virtual-row-offset'
 import type { Row } from './worktree-list-groups'
 import type { ProjectGroup } from '../../../../shared/types'
 
@@ -48,25 +49,12 @@ export function getSiblingGroupIdsByParent(rows: readonly Row[]): Map<string | n
       continue
     }
     seen.add(row.projectGroup.id)
-    const parent =
-      'parentGroupId' in row.projectGroup ? (row.projectGroup.parentGroupId ?? null) : null
+    const parent = row.projectGroup.parentGroupId ?? null
     const list = byParent.get(parent) ?? []
     list.push(row.projectGroup.id)
     byParent.set(parent, list)
   }
   return byParent
-}
-
-function getVirtualRowStart(virtualRow: HTMLElement | null): number | null {
-  if (!virtualRow) {
-    return null
-  }
-  const rawStart = virtualRow.getAttribute('data-worktree-virtual-row-start')
-  if (rawStart === null) {
-    return null
-  }
-  const start = Number(rawStart)
-  return Number.isFinite(start) ? start : null
 }
 
 export function measureGroupHeaderDragRects(
