+++ src/renderer/src/components/sidebar/project-header-drop.ts
@@ -1,6 +1,7 @@
 import { getEffectiveProjectGroupManualRank } from '../../../../shared/project-groups'
 import { interpolateSparseOrder } from './sidebar-drop-order-interpolation'
 import { getWorktreeSidebarBoundaryDrop } from './worktree-sidebar-drag-autoscroll'
+import { getVirtualRowStart } from './sidebar-virtual-row-offset'
 import type { Row } from './worktree-list-groups'
 import type { Repo } from '../../../../shared/types'
 
@@ -88,18 +89,6 @@ export function mapSidebarProjectHeaderDropIndexToSiblingInsertIndex(args: {
       ? args.sidebarDropIndex - 1
       : args.sidebarDropIndex
   return Math.max(0, Math.min(args.siblingCount, adjustedDropIndex))
-}
-
-function getVirtualRowStart(virtualRow: HTMLElement | null): number | null {
-  if (!virtualRow) {
-    return null
-  }
-  const rawStart = virtualRow.getAttribute('data-worktree-virtual-row-start')
-  if (rawStart === null) {
-    return null
-  }
-  const start = Number(rawStart)
-  return Number.isFinite(start) ? start : null
 }
 
 export function measureProjectHeaderDragRects(
