+++ src/renderer/src/components/sidebar/sidebar-virtual-row-offset.ts
@@ -1,0 +1,13 @@
+/** Resolve a virtualized sidebar row's absolute top from its
+ *  data-worktree-virtual-row-start attribute; null when unset/invalid. */
+export function getVirtualRowStart(virtualRow: HTMLElement | null): number | null {
+  if (!virtualRow) {
+    return null
+  }
+  const rawStart = virtualRow.getAttribute('data-worktree-virtual-row-start')
+  if (rawStart === null) {
+    return null
+  }
+  const start = Number(rawStart)
+  return Number.isFinite(start) ? start : null
+}
